+++ packages/amplify-graphql-api-construct-tests/src/__tests__/backends/amplify-table/simple-todo/updateKeySchema/enabled/app.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
-import { App, Stack, Duration } from 'aws-cdk-lib';
+import { App, Stack, Duration, Tags } from 'aws-cdk-lib';
 // @ts-ignore
 import { AmplifyGraphqlApi, AmplifyGraphqlDefinition } from '@aws-amplify/graphql-api-construct';
 
@@ -33,3 +33,7 @@ new AmplifyGraphqlApi(stack, 'GraphqlApi', {
     allowDestructiveGraphqlSchemaUpdates: true,
   },
 });
+
+Tags.of(stack).add('created-by', 'amplify-updated');
+Tags.of(stack).add('amplify:deployment-type', 'sandbox-updated');
+Tags.of(stack).add('amplify:friendly-name', 'amplifyData-updated');