+++ packages/amplify-graphql-relational-transformer/src/graphql-has-many-transformer.ts
@@ -17,6 +17,7 @@ import {
   TransformerSchemaVisitStepContextProvider,
   TransformerTransformSchemaStepContextProvider,
 } from '@aws-amplify/graphql-transformer-interfaces';
+import { Annotations } from 'aws-cdk-lib';
 import {
   DirectiveNode,
   DocumentNode,
@@ -25,6 +26,7 @@ import {
   NamedTypeNode,
   ObjectTypeDefinitionNode,
   ObjectTypeExtensionNode,
+  Kind,
 } from 'graphql';
 import { getBaseType, isListType, isNonNullType, makeField, makeNamedType, makeNonNullType } from 'graphql-transformer-common';
 import produce from 'immer';
@@ -149,6 +151,22 @@ export class HasManyTransformer extends TransformerPluginBase {
     const context = ctx as TransformerContextProvider;
 
     for (const config of this.directiveList) {
+      // This validation can't occur in validate because the api has not been initialized until generateResolvers
+      if (!ctx.transformParameters.allowGen1Patterns) {
+        const { field, object } = config;
+        const modelName = object.name.value;
+        const fieldName = field.name.value;
+        if (field.type.kind === Kind.NON_NULL_TYPE) {
+          Annotations.of(ctx.api).addWarning(
+            `@${HasManyDirective.name} on required fields is deprecated. Modify ${modelName}.${fieldName} to be optional. This functionality will be removed in the next major release.`,
+          );
+        }
+        if (config.fields) {
+          Annotations.of(ctx.api).addWarning(
+            `fields argument on @${HasManyDirective.name} is deprecated. Modify ${modelName}.${fieldName} to use references instead. This functionality will be removed in the next major release.`,
+          );
+        }
+      }
       const dbType = getStrategyDbTypeFromTypeNode(config.field.type, context);
       const dataSourceBasedTransformer = getHasManyDirectiveTransformer(dbType, config);
       dataSourceBasedTransformer.generateResolvers(ctx, config);
@@ -158,7 +176,6 @@ export class HasManyTransformer extends TransformerPluginBase {
 
 const validate = (config: HasManyDirectiveConfiguration, ctx: TransformerContextProvider): void => {
   const { field } = config;
-
   if (!isListType(field.type)) {
     throw new InvalidDirectiveError(`@${HasManyDirective.name} must be used with a list. Use @hasOne for non-list types.`);
   }