+++ packages/amplify-graphql-relational-transformer/src/graphql-has-one-transformer.ts
@@ -16,6 +16,7 @@ import {
   ModelDataSourceStrategyDbType,
 } from '@aws-amplify/graphql-transformer-interfaces';
 import { HasOneDirective } from '@aws-amplify/graphql-directives';
+import { Annotations } from 'aws-cdk-lib';
 import {
   ArgumentNode,
   DirectiveNode,
@@ -24,6 +25,7 @@ import {
   InterfaceTypeDefinitionNode,
   NamedTypeNode,
   ObjectTypeDefinitionNode,
+  Kind,
 } from 'graphql';
 import {
   getBaseType,
@@ -177,6 +179,22 @@ export class HasOneTransformer extends TransformerPluginBase {
     const context = ctx as TransformerContextProvider;
 
     for (const config of this.directiveList) {
+      // This validation can't occur in validate because the api has not been initialized until generateResolvers
+      if (!ctx.transformParameters.allowGen1Patterns) {
+        const { field, object } = config;
+        const modelName = object.name.value;
+        const fieldName = field.name.value;
+        if (field.type.kind === Kind.NON_NULL_TYPE) {
+          Annotations.of(ctx.api).addWarning(
+            `@${HasOneDirective.name} on required fields is deprecated. Modify ${modelName}.${fieldName} to be optional. This functionality will be removed in the next major release.`,
+          );
+        }
+        if (config.fields) {
+          Annotations.of(ctx.api).addWarning(
+            `fields argument on @${HasOneDirective.name} is deprecated. Modify ${modelName}.${fieldName} to use references instead. This functionality will be removed in the next major release.`,
+          );
+        }
+      }
       const dbType = getStrategyDbTypeFromTypeNode(config.field.type, context);
       const dataSourceBasedTransformer = getHasOneDirectiveTransformer(dbType, config);
       dataSourceBasedTransformer.generateResolvers(ctx, config);
@@ -186,7 +204,6 @@ export class HasOneTransformer extends TransformerPluginBase {
 
 const validate = (config: HasOneDirectiveConfiguration, ctx: TransformerContextProvider): void => {
   const { field } = config;
-
   let dbType: ModelDataSourceStrategyDbType;
   try {
     // getStrategyDbTypeFromTypeNode throws if a datasource is not found for the model. We want to catch that condition