+++ packages/amplify-graphql-transformer-core/src/index.ts
@@ -65,6 +65,7 @@ export {
   isMutationNode,
   isObjectTypeDefinitionNode,
   isQueryNode,
+  isQueryType,
   isSqlDbType,
   isSqlModel,
   isSqlStrategy,
@@ -73,6 +74,13 @@ export {
   SQLLambdaResourceNames,
 } from './utils';
 export type { SetResourceNameProps } from './utils';
+export {
+  type JSONSchema,
+  isDisallowedScalarType,
+  supportedScalarTypes,
+  GraphQLScalarJSONSchemaDefinition,
+  convertNamedTypeToJSONSchema,
+} from './utils/ai';
 export * from './utils/operation-names';
 export * from './errors';
 export {