+++ packages/amplify-graphql-index-transformer/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@3.0.1...@aws-amplify/graphql-index-transformer@3.0.2) (2024-09-06)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
+## [3.0.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@3.0.0...@aws-amplify/graphql-index-transformer@3.0.1) (2024-08-28)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
+# [3.0.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.9...@aws-amplify/graphql-index-transformer@3.0.0) (2024-08-20)
+
+- feat!: bump version for v2 transformer packages ([7dac35c](https://github.com/aws-amplify/amplify-category-api/commit/7dac35cceb971b256b5688b0745ef82afc78b641))
+
+### BREAKING CHANGES
+
+- distinguish from LTS version
+
+## [2.4.9](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.8...@aws-amplify/graphql-index-transformer@2.4.9) (2024-08-12)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
+## [2.4.8](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.7...@aws-amplify/graphql-index-transformer@2.4.8) (2024-07-25)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
+## [2.4.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.6...@aws-amplify/graphql-index-transformer@2.4.7) (2024-07-15)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
+## [2.4.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.5...@aws-amplify/graphql-index-transformer@2.4.6) (2024-07-02)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
+## [2.4.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.4...@aws-amplify/graphql-index-transformer@2.4.5) (2024-07-01)
+
+### Bug Fixes
+
+- auth to use validateUsingSource in place of auth filter to show error message ([#2523](https://github.com/aws-amplify/amplify-category-api/issues/2523)) ([b7d83f9](https://github.com/aws-amplify/amplify-category-api/commit/b7d83f991f85eaffb2408cff98e1880c7fa680ef))
+
+## [2.4.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.3...@aws-amplify/graphql-index-transformer@2.4.4) (2024-06-25)
+
+**Note:** Version bump only for package @aws-amplify/graphql-index-transformer
+
 ## [2.4.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-index-transformer@2.4.2...@aws-amplify/graphql-index-transformer@2.4.3) (2024-05-15)
 
 **Note:** Version bump only for package @aws-amplify/graphql-index-transformer